image: cast ELF scatter log fields to unsigned long
offset/filesz are uint64_t now; cast them for the %08lx/%lu
conversions so the varargs read matches on 32-bit targets, matching
the check function and the rejection messages above.

+++ src/image.c
@@ -2555,7 +2555,8 @@ int wolfBoot_load_flash_image_elf(int part, unsigned long* entry_out, int ext_fl
 
         wolfBoot_printf("ELF: [STORE] Writing loadable segment: "
                         "loadaddr=0x%08lx, offset=0x%08lx, size=%lu\n",
-                        (unsigned long)load_addr, offset, filesz);
+                        (unsigned long)load_addr, (unsigned long)offset,
+                        (unsigned long)filesz);
         if (copy_flash_buffered((uintptr_t)(image + offset), load_addr,
                                 filesz, ext_flash, ext_flash) != 0) {
             wolfBoot_printf("ELF: [STORE] ERROR: could not write "
